+++ context_di/example/lib/features/basic/presentation/basic_feature.dart
@@ -6,7 +6,7 @@ import '../domain/domain.dart';
 import 'managers/entity_bloc.dart';
 import 'managers/list_bloc.dart';
 
-part 'basic.feature.dart';
+part 'basic_feature.g.dart';
 
 typedef EntityBlocParams = ({int id});
 
@@ -16,6 +16,6 @@ void _dispose(BuildContext context, Repository instance) => instance.dispose();
 @Singleton(Repository, as: RepositoryInterface, dispose: _dispose)
 @Factory(ListBloc)
 @Factory(EntityBloc, params: EntityBlocParams)
-class Basic extends FeatureDependencies with _$BasicFeatureMixin {
-  const Basic({super.key, super.builder});
+class BasicFeature extends FeatureDependencies with _$BasicFeatureMixin {
+  const BasicFeature({super.key, super.builder, super.child});
 }