+++ context_di/example/lib/features/features.dart
@@ -2,7 +2,7 @@
 export 'basic/data/repositories/repository.dart';
 export './basic/domain/entities/some_entity.dart';
 export 'basic/domain/repositories/repository_interface.dart';
-export 'basic/presentation/basic.dart';
+export 'basic/presentation/basic_feature.dart';
 export 'basic/presentation/basic_feature_page.dart';
 export 'basic/presentation/managers/entity_bloc.dart';
 export 'basic/presentation/pages/entity_page.dart';