+++ context_di/lib/utils.dart
@@ -1,0 +1,4 @@
+import 'package:flutter/widgets.dart';
+
+typedef Factory<T> = T Function(BuildContext context);
+typedef ParamsFactory<T, P> = T Function(BuildContext context, P params);